requirements.txt
requests==2.32.4
lxml==6.0.0
selenium==4.34.2
pillow
# oss2~=2.15.0
webdriver-manager==4.0.2